@@ -14,10 +14,14 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  # sdk: ">=2.6.0 <3.0.0"
+  # sdk: ">=2.2.2 <3.0.0"
+  sdk: ">=2.7.0 <3.0.0"
 
 dependencies:
   flutter:
+    sdk: flutter
+  flutter_localizations:    
     sdk: flutter
 
   # The following adds the Cupertino Icons font to your application.
@@ -35,19 +39,33 @@ dependencies:
   shared_preferences: ^0.5.6+3
   flutter_slidable: ^0.5.4
   image_crop: ^0.3.2
-  # flutter_screenutil: ^1.1.0
-  # dash_chat: ^1.0.18
   connectivity: ^0.4.3+1
-  provider: ^2.0.1
+  provider: ^4.1.3
   data_connection_checker: ^0.3.0
   animated_size_and_fade: ^1.1.0
   flutter_spinkit: ^4.1.2
   flutter_local_notifications: ^1.4.1
   rxdart: ^0.24.0
-  url_launcher: ^5.4.5
   html: ^0.14.0+3
   cached_network_image: ^2.1.0+1
   flutter_svg: ^0.17.4
+  network_image_to_byte: ^0.0.1
+  flutter_webview_plugin: ^0.3.11
+  collection_ext: ^0.2.0
+  tuple: ^1.0.3
+  url_launcher: ^5.4.2  
+  sliding_up_panel: ^1.0.2
+  flutter_tagging: ^2.2.0+3
+  # reorderables: ^0.3.2
+  pull_to_refresh: ^1.5.7
+  bot_toast: ^3.0.0
+  flutter_datetime_picker: ^1.3.8
+  flutter_app_badger: ^1.1.2
+  badges: ^1.1.1
+  flutter_screenutil: ^2.1.0
+  flutter_staggered_grid_view: ^0.3.0
+  after_layout: ^1.0.7+2
+
 
 dev_dependencies:
   flutter_test:
@@ -70,10 +88,12 @@ flutter:
   #  - images/a_dot_burr.jpeg
   #  - images/a_dot_ham.jpeg
   assets:
-    - assets/images/user_logo.jpeg
-    - assets/images/login_logo.jpg
-    - assets/images/nav_me.jpg
-    - assets/images/screen2.jpg
+    - assets/images/
+    
+  fonts:
+    - family: app-icons
+      fonts:
+        - asset: assets/fonts/app-icons.ttf
 
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware.
