@@ -1,42 +1,39 @@
+$(function() {
+    //make connection
+    var socket = io.connect('http://localhost:3000/')
 
-$(function(){
-   	//make connection
-	var socket = io.connect('http://localhost:3000/')
+    //buttons and inputs
+    var message = $("#message")
+    var username = $("#username")
+    var send_message = $("#send_message")
+    var send_username = $("#send_username")
+    var chatroom = $("#chatroom")
+    var feedback = $("#feedback")
 
-	//buttons and inputs
-	var message = $("#message")
-	var username = $("#username")
-	var send_message = $("#send_message")
-	var send_username = $("#send_username")
-	var chatroom = $("#chatroom")
-	var feedback = $("#feedback")
+    //Emit message
+    send_message.click(function() {
+        socket.emit('chat', { message: message.val() })
+    })
 
-	//Emit message
-	send_message.click(function(){
-		socket.emit('send_message', {message : message.val()})
-	})
+    //Listen on new_message
+    socket.on("chat", (data) => {
+        feedback.html('');
+        message.val('');
+        chatroom.append("<p class='message'>" + data.username + ": " + data.message + "</p>")
+    })
 
-	//Listen on new_message
-	socket.on("new_message", (data) => {
-		feedback.html('');
-		message.val('');
-		chatroom.append("<p class='message'>" + data.username + ": " + data.message + "</p>")
-	})
+    //Emit a username
+    send_username.click(function() {
+        socket.emit('change_username', { username: username.val() })
+    })
 
-	//Emit a username
-	send_username.click(function(){
-		socket.emit('change_username', {username : username.val()})
-	})
+    //Emit typing
+    message.bind("keypress", () => {
+        socket.emit('typing')
+    })
 
-	//Emit typing
-	message.bind("keypress", () => {
-		socket.emit('typing')
-	})
-
-	//Listen on typing
-	socket.on('typing', (data) => {
-		feedback.html("<p><i>" + data.username + " is typing a message..." + "</i></p>")
-	})
+    //Listen on typing
+    socket.on('typing', (data) => {
+        feedback.html("<p><i>" + data.username + " is typing a message..." + "</i></p>")
+    })
 });
-
-
