@@ -1,23 +1,24 @@
 import 'dart:async';
 
 import 'package:flutter/services.dart';
+import 'package:keep/utils/utils_class.dart' show ClipBoardData;
 
 class ClipboardStream {
-  static final clipboardContentStream = StreamController<String>.broadcast();
+  static final clipboardContentStream =
+      StreamController<ClipBoardData>.broadcast();
 
   final Timer clipboardTriggerTime = Timer.periodic(
-    const Duration(seconds: 2),
+    const Duration(seconds: 1),
     (timer) {
       Clipboard.getData('text/plain').then((clipboarContent) {
         String url = clipboarContent != null ? clipboarContent.text : '';
         print('Clipboard content ${url.length}');
-
-        clipboardContentStream.add(url);
+        clipboardContentStream.add(ClipBoardData(url));
       });
     },
   );
 
-  Stream get clipboardText => clipboardContentStream.stream;
+  Stream<ClipBoardData> get clipboardText => clipboardContentStream.stream;
 
   clear() async {
     await Clipboard.setData(ClipboardData(text: ''));
@@ -25,7 +26,6 @@ class ClipboardStream {
 
   void dispose() {
     clipboardContentStream.close();
-
     clipboardTriggerTime.cancel();
   }
 }
