@@ -5,12 +5,15 @@ import 'package:flutter/material.dart';
 import 'package:image_picker/image_picker.dart';
 import 'package:keep/UI/Home/Settings/uploadApi/upload_page.dart';
 import 'package:keep/data/provider/user_provider.dart';
-import 'package:keep/global/flush_status.dart';
-import 'package:keep/global/global_tool.dart';
+import 'package:keep/widget/flush_status.dart';
+import 'package:keep/widget/component_widget.dart';
 import 'package:provider/provider.dart';
+import 'package:keep/utils/utils_class.dart';
 import 'settingUI/user_panel.dart';
 import 'settingUI/user_info.dart';
 import 'settingUI/setting_item.dart';
+import 'package:keep/settings/selection_config.dart';
+import 'package:keep/settings/status_config.dart';
 
 class SettingMainPage extends StatefulWidget {
   @override
@@ -23,11 +26,13 @@ class SettingMainPageState extends State<SettingMainPage> {
   BuildContext _ctx;
 
   Future<File> chooseGallery() async {
+    // ignore: deprecated_member_use
     File img = await ImagePicker.pickImage(source: ImageSource.gallery);
     return img;
   }
 
   Future<File> chooseCamara() async {
+    // ignore: deprecated_member_use
     File img = await ImagePicker.pickImage(source: ImageSource.camera);
     return img;
   }
@@ -40,11 +45,11 @@ class SettingMainPageState extends State<SettingMainPage> {
         print('upload failed');
       } else {
         // 局部刷新界面
-        _key.currentState.setAvatar(res.stream['avatar']);
-        Provider.of<UserProvider>(context, listen: true)
+        Provider.of<UserProvider>(context, listen: false)
             .updateUserPic(res.stream['avatar']);
-            
-        print('success.....');
+        _key.currentState.setAvatar(UserProvider().userAvatar);
+
+        // print('success.....');
         showFlushBar(context, username, res.stream['hint_msg'],
             iconIndicator['success']);
       }
@@ -57,21 +62,29 @@ class SettingMainPageState extends State<SettingMainPage> {
     return Consumer<UserProvider>(builder: (context, user, child) {
       return Scaffold(
         body: Stack(children: [
-          Container(
-              child: ListView(
-            physics: NeverScrollableScrollPhysics(),
-            shrinkWrap: true,
+          ListView(
+            // physics: NeverScrollableScrollPhysics(),
+            // shrinkWrap: true,
             children: <Widget>[
-              buildHeader(),
-              UserPanel(
+              Container(
+                  color: Colors.blueGrey,
+                  child: Column(children: [
+                    buildHeader(),
+                    UserPanel(
+                        username: user.username,
+                        email: user.email,
+                        avatar: user.userAvatar,
+                        key: _key),
+                  ])),
+              UserInfo(
                   username: user.username,
                   email: user.email,
-                  avatar: user.avatar,
-                  key: _key),
-              UserInfo(username: user.username, email: user.email, bio: null),
+                  phone: user.phone,
+                  bio: null),
+              buildGap(),
               SettingItems()
             ],
-          )),
+          ),
           buildButton(user.username),
         ]),
       );
@@ -86,8 +99,8 @@ class SettingMainPageState extends State<SettingMainPage> {
             mainAxisSize: MainAxisSize.min,
             children: <Widget>[
               new ListTile(
-                leading: new Icon(Icons.photo_camera),
-                title: new Text("Camera"),
+                leading: const Icon(Icons.photo_camera),
+                title: const Text("Camera"),
                 onTap: () async {
                   chooseCamara().then((file) {
                     if (file != null) {
@@ -98,8 +111,8 @@ class SettingMainPageState extends State<SettingMainPage> {
                 },
               ),
               new ListTile(
-                leading: new Icon(Icons.photo_library),
-                title: new Text("Gallery"),
+                leading: const Icon(Icons.photo_library),
+                title: const Text("Gallery"),
                 onTap: () async {
                   chooseGallery().then((file) {
                     if (file != null) {
@@ -115,28 +128,31 @@ class SettingMainPageState extends State<SettingMainPage> {
   }
 
   Widget buildButton(String username) {
-    return Container(
-        decoration: BoxDecoration(
-          shape: BoxShape.circle,
-          border: Border.all(width: 1.0, color: Colors.grey.withAlpha(20)),
-          boxShadow: [
-            BoxShadow(
-              color: Colors.grey,
-              blurRadius: 5.0,
-              offset: Offset(0.3, 0.3),
-            )
-          ],
-        ),
-        margin: EdgeInsets.only(top: 150.0, left: 320.0),
-        child: CircleAvatar(
-          radius: 28.0,
-          backgroundColor: Colors.white,
-          child: Container(
-              child: IconButton(
-                  icon:
-                      Icon(Icons.photo_camera, color: Colors.grey, size: 28.0),
-                  onPressed: () => showBottomSheet(context, username))),
-        ));
+    return Positioned(
+        top: 150.0,
+        left: MediaQuery.of(_ctx).size.width * 0.75,
+        child: Container(
+            decoration: BoxDecoration(
+              shape: BoxShape.circle,
+              border: Border.all(width: 1.0, color: Colors.grey.withAlpha(20)),
+              boxShadow: [
+                BoxShadow(
+                  color: Colors.grey,
+                  blurRadius: 5.0,
+                  offset: Offset(0.3, 0.3),
+                )
+              ],
+            ),
+            // margin: EdgeInsets.only(top: 150.0, left: 320.0),
+            child: CircleAvatar(
+              radius: 28.0,
+              backgroundColor: Colors.white,
+              child: Container(
+                  child: IconButton(
+                      icon: const Icon(Icons.photo_camera,
+                          color: Colors.grey, size: 28.0),
+                      onPressed: () => showBottomSheet(context, username))),
+            )));
   }
 
   Widget buildHeader() {
@@ -148,7 +164,7 @@ class SettingMainPageState extends State<SettingMainPage> {
           Flexible(
               flex: 3,
               child: IconButton(
-                  icon: Icon(Icons.arrow_back),
+                  icon: const Icon(Icons.arrow_back),
                   onPressed: () => Navigator.pop(context))),
           Flexible(
               flex: 10,
@@ -164,7 +180,7 @@ class SettingMainPageState extends State<SettingMainPage> {
                     child: DropdownButton<String>(
                   // value: dropdownValue,
                   iconEnabledColor: Color(0xFF595959),
-                  icon: Icon(
+                  icon: const Icon(
                     Icons.more_vert,
                     color: Colors.black,
                   ),
@@ -178,7 +194,7 @@ class SettingMainPageState extends State<SettingMainPage> {
                       dropdownValue = newValue;
                     });
                   },
-                  items: <String>[
+                  items: const <String>[
                     'Edit name',
                     'Log out',
                   ].map<DropdownMenuItem<String>>((String value) {
