@@ -1,286 +1,235 @@
-import 'dart:async';
-import 'dart:io';
-import 'dart:typed_data';
 import 'dart:ui';
 
-// import 'package:cached_network_image/cached_network_image.dart';
+import 'package:badges/badges.dart';
+import 'package:provider/provider.dart';
+import 'package:after_layout/after_layout.dart';
+import 'package:adhara_socket_io/adhara_socket_io.dart';
+
+import 'package:flutter/services.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/rendering.dart';
-import 'package:flutter/services.dart';
-import 'package:adhara_socket_io/adhara_socket_io.dart';
-import 'package:flutter_local_notifications/flutter_local_notifications.dart';
-import 'package:flutter_svg/flutter_svg.dart';
-import 'package:http/http.dart' as http;
-import 'package:html/parser.dart';
-import 'package:keep/BLoC/clipboard_listener.dart';
-import 'package:keep/data/rest_ds.dart';
-import 'package:path_provider/path_provider.dart';
+// import 'package:flutter_local_notifications/flutter_local_notifications.dart';
 
-import 'package:keep/UI/Home/chat/messages_page.dart';
-import 'package:keep/models/friend.dart';
+import 'package:keep/service/socket_util.dart';
+
+import 'package:keep/data/sputil.dart';
+import 'package:keep/data/provider/badgeNum_provider.dart';
+import 'package:keep/data/provider/user_provider.dart';
 import 'package:keep/data/provider/friend_provider.dart';
-import 'package:keep/global/global_tool.dart';
+import 'package:keep/data/repository/message_repository.dart';
+import 'package:keep/BLoC/message_bloc.dart';
+
+import 'package:keep/models/friend.dart';
+
 import 'package:keep/utils/event_util.dart';
-import 'package:keep/utils/socket_util.dart';
-import 'package:keep/global/notification.dart';
+import 'package:keep/utils/tools_function.dart';
+
+import 'package:keep/settings/notification_helper.dart';
+
 import 'nav_drawer.dart';
+import 'package:keep/widget/over_scroll.dart';
+import 'package:keep/widget/storage_panel_widget.dart';
+import 'package:keep/UI/Home/Chat/ChatScreenUI/chatscreen.dart';
+import 'package:keep/UI/Home/chat/messages_page.dart';
+import 'package:keep/UI/Home/Noting/note_page.dart';
+import 'package:keep/UI/Home/Todo/todo_page.dart' show TodoPage;
 
 class MyHomePage extends StatefulWidget {
   @override
   _MyHomePageState createState() => _MyHomePageState();
 }
 
-class _MyHomePageState extends State<MyHomePage> {
+class _MyHomePageState extends State<MyHomePage>
+    with SingleTickerProviderStateMixin, AfterLayoutMixin<MyHomePage> {
   final _kTabPages = <Widget>[
-    Center(child: Icon(Icons.cloud, size: 64.0, color: Colors.teal)),
-    Center(child: Icon(Icons.alarm, size: 64.0, color: Colors.cyan)),
-    // buildAvatar(_imgUrl),
-    MessageView(),
+    Center(child: NotePage()),
+    Center(child: TodoPage()),
+    Center(child: MessageView()),
   ];
-  final _kTabs = <Tab>[
-    Tab(icon: Icon(Icons.cloud), text: 'Fragmentation'),
-    Tab(icon: Icon(Icons.alarm), text: 'Todo'),
-    Tab(icon: Icon(Icons.forum), text: 'User'),
-  ];
-  final bloc = new ClipboardStream();
+  TabController _tabController;
+  final String username = UserProvider().username;
+  final Object avatar = UserProvider().userAvatar;
+  final userId = UserProvider.getUserId();
+  final apiKey = UserProvider.getApiKey();
+  final msgRepo = new MessageRepository();
+  final msgBloc = new MessageBloc();
 
-  // Network Banner Status
-  var connectionStatus;
-  String status;
-  String _imgUrl = 'https://static.zhihu.com/heifetz/assets/logo.f6eef033.png';
+  final _textController = new TextEditingController();
 
+  // notification
+  // FlutterLocalNotificationsPlugin flutterLocalNotificationsPlugin;
   final MethodChannel platform =
       MethodChannel('crossingthestreams.io/resourceResolver');
 
-  @override
-  void dispose() {
-    didReceiveLocalNotificationSubject.close();
-    selectNotificationSubject.close();
-    super.dispose();
-  }
+  BuildContext _ctx;
 
+  static const textStyle = TextStyle(
+      fontSize: 12.0,
+      color: Colors.white,
+      fontFamily: 'OpenSans',
+      fontWeight: FontWeight.w600);
   SocketIO _socket;
 
   initSocket() async {
     return await new SocketUtil().socket;
   }
 
+  void msgStreamCallback() => msgBloc.getMessages();
+
   @override
   void initState() {
     super.initState();
-    _requestIOSPermissions();
-    _configureDidReceiveLocalNotificationSubject();
+
+    int initialIndex = SpUtil.getInt('$userId-initialIndex', defValue: 1);
+    _tabController =
+        new TabController(length: 3, vsync: this, initialIndex: initialIndex);
+    _tabController.addListener(
+        () => SpUtil.putInt('$userId-initialIndex', _tabController.index));
+    // _requestIOSPermissions();
+    // _configNotification();
     _configureSelectNotificationSubject();
-    initSocket().then((socket) => _socket = socket);
-    bus.on('new_msg_notification', (msg) async {
-      await _showNotification();
-      // await _showNotificationWithNoBody();
-      // await _cancelNotification();
-      // await _scheduleNotification();
-    });
-    bus.on('friendRequest', (data) async {
-      // print(data);
-      await _showNotification();
-      // 此处默认接受好友请求，后续需要开发好友请求显示页面
-      _socket.emit('friendResponse', [
-        {
-          'status': 1,
-          'userOneId': data['userOneId'],
-          'userTwoId': data['userTwoId']
-        }
-      ]);
-    });
-    bus.on('friendResponse', (data) async {
-      print(data);
-      await _showNotification();
-    });
+    _configureDidReceiveLocalNotificationSubject();
+  }
 
-    bus.on('friendDone', (data) {
-      FriendProvider.addFriend(Friend.fromMap(data));
-    });
+  void onChat(stream) {
+    print('listen...');
+    stream['context'] = context;
+    onChatCallback(stream);
+  }
 
-    bus.on('clipboard_url', (url) {
-      print('trigger the save_url');
-      RestDatasource _api = new RestDatasource();
-      _api
-          .getUrlContent(url,
-              headers: {
-                'Accept': '*/*',
-                'Accept-Language': 'en-US,en;q=0.8',
-                'Cache-Control': 'max-age=0',
-                'User-Agent':
-                    'Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/48.0.2564.116 Safari/537.36',
-                'Connection': 'keep-alive',
-                'Referer': 'http://www.baidu.com/'
-              },
-              isJson: false)
-          .then((res) {
-        var document = parse(res);
-        print(document.getElementsByTagName('title')[0].text);
-        setState(() {
-          var imgs = document.getElementsByTagName('img');
-          print(imgs[0].attributes['src']);
-          // imgs.forEach((img)
-          for (final img in imgs) {
-            print(img.attributes['src']);
-            if (img.attributes['src'] != null &&
-                img.attributes['src'].endsWith('svg')) {
-              _imgUrl = img.attributes['src'];
-              break;
-            } else if (img.attributes['src'] != null &&
-                (img.attributes['src'].endsWith('jpg') ||
-                    img.attributes['src'].endsWith('jpeg') ||
-                    img.attributes['src'].endsWith('png')) &&
-                img.attributes['src'] !=
-                    'https://static.zhihu.com/heifetz/assets/logo.f6eef033.png') {
-              _imgUrl = img.attributes['src'];
-              break;
-            } else {
-              _imgUrl = '';
-            }
-          }
-          print(_imgUrl);
-        });
-      });
+  void onChats(stream) {
+    stream['context'] = context;
+    onChatsCallback(stream);
+  }
+
+  void onFriendRequest(stream) {
+    stream['context'] = context;
+    onFriendRequestCallback(stream);
+  }
+
+  void onFriendsRequest(stream) {
+    stream['context'] = context;
+    onFriendsRequestCallback(stream);
+  }
+
+  @override
+  void afterFirstLayout(BuildContext context) {
+    // Calling the same function "after layout" to resolve the issue.
+    initSocket().then((socket) {
+      _socket = socket;
+      _socket.on('chat', onChat);
+      _socket.on('chats', onChats);
+      _socket.on('friendRequest', onFriendRequest);
+      _socket.on('friendRequests', onFriendsRequest);
+      _socket.connect();
     });
+  }
+
+  @override
+  void dispose() {
+    _socket.off('chat');
+    _socket.off('chats');
+    _socket.off('friendRequest');
+    _socket.off('friendRequests');
+    bus.off('scratchAllMessages', (_) => msgStreamCallback);
+    _tabController.dispose();
+    flutterLocalNotificationsPlugin.cancelAll();
+    didReceiveLocalNotificationSubject.close();
+    selectNotificationSubject.close();
+
+    msgBloc.dispose();
+    super.dispose();
   }
 
   @override
   Widget build(BuildContext context) {
+    debugPrint('loading tab view page....');
+    _ctx = context;
     precacheImage(AssetImage("assets/images/nav_me.jpg"), context);
+    bus.on('scratchAllMessages', (_) => msgStreamCallback);
+    // bus.on('triggerMsgNotification', (stream) => onChatCallback(stream));
     return Material(
+        child: Stack(children: [
+      buildTabView(),
+      StoragePanel(textEditingController: _textController)
+    ]));
+  }
+
+  Widget buildTabView() {
+    return ChangeNotifierProvider(
+        create: (context) => BadgeNumProvider(),
         child: DefaultTabController(
-            length: _kTabs.length,
+            length: _kTabPages.length,
             child: Scaffold(
-                drawer: NavDrawer(),
-                appBar: AppBar(
-                  title: Text('Keep'),
-                  backgroundColor: Colors.cyan,
-                  bottom: TabBar(
-                    tabs: _kTabs,
-                  ),
+              resizeToAvoidBottomPadding: false,
+              drawer: NavDrawer(parentContext: _ctx),
+              appBar: AppBar(
+                title: Text('Keep'),
+                backgroundColor: Colors.cyan,
+                bottom: TabBar(
+                  controller: _tabController,
+                  tabs: <Tab>[
+                    Tab(icon: const Icon(Icons.cloud), text: 'Fragmentation'),
+                    Tab(icon: const Icon(Icons.alarm), text: 'Todo'),
+                    Tab(
+                        icon: Consumer<BadgeNumProvider>(
+                          builder: (context, badge, child) => badge
+                                      .badgeNumber >
+                                  0
+                              ? GestureDetector(
+                                  onHorizontalDragCancel: () =>
+                                      Provider.of<BadgeNumProvider>(context,
+                                              listen: false)
+                                          .clear(),
+                                  onVerticalDragCancel: () =>
+                                      Provider.of<BadgeNumProvider>(context,
+                                              listen: false)
+                                          .clear(),
+                                  child: Badge(
+                                      position: BadgePosition.bottomLeft(
+                                          bottom: 12, left: 12),
+                                      // padding: EdgeInsets.all(2.0),
+                                      animationDuration:
+                                          Duration(milliseconds: 300),
+                                      animationType: BadgeAnimationType.fade,
+                                      badgeColor:
+                                          Colors.redAccent.withAlpha(225),
+                                      badgeContent: Text(
+                                          badge.badgeNumber > 99
+                                              ? '99+'
+                                              : badge.badgeNumber.toString(),
+                                          style: TextStyle(
+                                              color: Colors.white,
+                                              fontSize: 12.0)),
+                                      child: child),
+                                )
+                              : child,
+                          child: const Icon(Icons.forum),
+                        ),
+                        text: 'User'),
+                  ],
+                  labelColor: Color(0xFF343434),
+                  labelStyle: textStyle.copyWith(
+                      fontSize: 14.0,
+                      color: Color(0xFFc9c9c9),
+                      fontWeight: FontWeight.w700),
+                  // unselectedLabelColor: Color(0xFFc9c9c9),
+                  unselectedLabelColor: Colors.white,
+                  unselectedLabelStyle: textStyle.copyWith(
+                      fontSize: 14.0,
+                      // color: Color(0xFFc9c9c9),
+                      color: Colors.white,
+                      fontWeight: FontWeight.w500),
                 ),
-                body: Stack(children: [
-                  TabBarView(
-                    children: _kTabPages,
-                  ),
-                  buildUrlPanel(context),
-                ]))));
-  }
-
-  buildAvatar(String imgUrl) {
-    if (imgUrl == null || imgUrl == '') {
-      return CircleAvatar(
-        child: Text('hello'),
-        backgroundColor: Colors.limeAccent,
-      );
-    } else if (imgUrl.endsWith('svg')) {
-      return new SvgPicture.network(
-        imgUrl,
-        height: 100.0,
-        width: 100.0,
-        // allowDrawingOutsideViewBox: true,
-      );
-    } else {
-      return Image.network(
-        imgUrl,
-        height: 100.0,
-        width: 100.0,
-      );
-      // return CachedNetworkImage(
-      //   imageUrl: imgUrl,
-      //   placeholder: (context, url) => new CircularProgressIndicator(),
-      //   errorWidget: (context, url, error) => new Icon(Icons.error),
-      // );
-    }
-  }
-
-  Widget buildUrlPanel(BuildContext context) {
-    return StreamBuilder<String>(
-      stream: bloc.clipboardText,
-      builder: (BuildContext context, AsyncSnapshot<String> snapshot) {
-        if (snapshot.hasData &&
-            snapshot.data.length > 0 &&
-            judgeUrl(snapshot.data)) {
-          return Positioned(
-            bottom: 0.0,
-            child: Container(
-              padding: EdgeInsets.symmetric(horizontal: 15.0),
-              color: Colors.greenAccent,
-              width: MediaQuery.of(context).size.width,
-              height: 90.0,
-              child: Row(
-                // mainAxisAlignment: MainAxisAlignment.spaceAround,
-                crossAxisAlignment: CrossAxisAlignment.center,
-                children: [
-                  Flexible(
-                      flex: 2,
-                      fit: FlexFit.tight,
-                      child: IconButton(
-                          icon: Icon(Icons.close),
-                          onPressed: () => bloc.clear())),
-                  Flexible(
-                      flex: 7,
-                      fit: FlexFit.tight,
-                      child: Container(
-                          padding: EdgeInsets.symmetric(vertical: 20.0),
-                          child: Column(
-                            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                            crossAxisAlignment: CrossAxisAlignment.start,
-                            children: [
-                              Text(
-                                'Save Copied Url?',
-                                style: TextStyle(
-                                  fontSize: 16.0,
-                                  fontFamily: 'Roboto',
-                                ),
-                              ),
-                              Text(
-                                snapshot.data,
-                                style: TextStyle(fontSize: 14.0),
-                                overflow: TextOverflow.ellipsis,
-                              )
-                            ],
-                          ))),
-                  Flexible(
-                      flex: 2,
-                      fit: FlexFit.tight,
-                      child: Container(
-                          width: MediaQuery.of(context).size.width * 0.25,
-                          child: FlatButton(
-                              onPressed: () {
-                                print('save the url ' + snapshot.data);
-                                bus.emit('clipboard_url', snapshot.data);
-                                bloc.clear();
-                              },
-                              child: Text(
-                                'Save',
-                                style: TextStyle(
-                                  fontSize: 16.0,
-                                  fontFamily: 'Roboto',
-                                ),
-                              ))))
-                ],
               ),
-            ),
-          );
-        } else {
-          return Container();
-        }
-      },
-    );
-  }
-
-  void _requestIOSPermissions() {
-    flutterLocalNotificationsPlugin
-        .resolvePlatformSpecificImplementation<
-            IOSFlutterLocalNotificationsPlugin>()
-        ?.requestPermissions(
-          alert: true,
-          badge: true,
-          sound: true,
-        );
+              body: ScrollConfiguration(
+                  behavior: OverScrollBehavior(),
+                  child: TabBarView(
+                    controller: _tabController,
+                    children: _kTabPages,
+                  )),
+            )));
   }
 
   void _configureDidReceiveLocalNotificationSubject() {
@@ -301,13 +250,13 @@ class _MyHomePageState extends State<MyHomePage> {
               child: Text('Ok'),
               onPressed: () async {
                 Navigator.of(context, rootNavigator: true).pop();
-                await Navigator.push(
-                  context,
-                  MaterialPageRoute(
-                    builder: (context) =>
-                        SecondScreen(receivedNotification.payload),
-                  ),
-                );
+                // await Navigator.push(
+                //   context,
+                //   MaterialPageRoute(
+                //     builder: (context) =>
+                //         SecondScreen(receivedNotification.payload),
+                //   ),
+                // );
               },
             )
           ],
@@ -317,627 +266,55 @@ class _MyHomePageState extends State<MyHomePage> {
   }
 
   void _configureSelectNotificationSubject() {
-    selectNotificationSubject.stream.listen((String payload) async {
-      await Navigator.push(
-        context,
-        MaterialPageRoute(builder: (context) => SecondScreen(payload)),
-      );
-    });
+    selectNotificationSubject.stream.listen(tapAction);
   }
 
-  Future<void> _showNotification() async {
-    var androidPlatformChannelSpecifics = AndroidNotificationDetails(
-        'your channel id', 'your channel name', 'your channel description',
-        importance: Importance.Max, priority: Priority.High, ticker: 'ticker');
-    var iOSPlatformChannelSpecifics = IOSNotificationDetails();
-    var platformChannelSpecifics = NotificationDetails(
-        androidPlatformChannelSpecifics, iOSPlatformChannelSpecifics);
-    await flutterLocalNotificationsPlugin.show(
-        0, 'plain title', 'plain body', platformChannelSpecifics,
-        payload: 'item x');
-  }
-
-  Future<void> _showNotificationWithNoBody() async {
-    var androidPlatformChannelSpecifics = AndroidNotificationDetails(
-        'your channel id', 'your channel name', 'your channel description',
-        importance: Importance.Max, priority: Priority.High, ticker: 'ticker');
-    var iOSPlatformChannelSpecifics = IOSNotificationDetails();
-    var platformChannelSpecifics = NotificationDetails(
-        androidPlatformChannelSpecifics, iOSPlatformChannelSpecifics);
-    await flutterLocalNotificationsPlugin.show(
-        0, 'plain title', null, platformChannelSpecifics,
-        payload: 'item x');
-  }
-
-  Future<void> _cancelNotification() async {
-    await flutterLocalNotificationsPlugin.cancel(0);
-  }
-
-  /// Schedules a notification that specifies a different icon, sound and vibration pattern
-  Future<void> _scheduleNotification() async {
-    var scheduledNotificationDateTime =
-        DateTime.now().add(Duration(seconds: 5));
-    var vibrationPattern = Int64List(4);
-    vibrationPattern[0] = 0;
-    vibrationPattern[1] = 1000;
-    vibrationPattern[2] = 5000;
-    vibrationPattern[3] = 2000;
-
-    var androidPlatformChannelSpecifics = AndroidNotificationDetails(
-        'your other channel id',
-        'your other channel name',
-        'your other channel description',
-        icon: 'secondary_icon',
-        sound: RawResourceAndroidNotificationSound('slow_spring_board'),
-        largeIcon: DrawableResourceAndroidBitmap('sample_large_icon'),
-        vibrationPattern: vibrationPattern,
-        enableLights: true,
-        color: const Color.fromARGB(255, 255, 0, 0),
-        ledColor: const Color.fromARGB(255, 255, 0, 0),
-        ledOnMs: 1000,
-        ledOffMs: 500);
-    var iOSPlatformChannelSpecifics =
-        IOSNotificationDetails(sound: 'slow_spring_board.aiff');
-    var platformChannelSpecifics = NotificationDetails(
-        androidPlatformChannelSpecifics, iOSPlatformChannelSpecifics);
-    await flutterLocalNotificationsPlugin.schedule(
-        0,
-        'scheduled title',
-        'scheduled body',
-        scheduledNotificationDateTime,
-        platformChannelSpecifics);
-  }
-
-  Future<void> _showNotificationWithNoSound() async {
-    var androidPlatformChannelSpecifics = AndroidNotificationDetails(
-        'silent channel id',
-        'silent channel name',
-        'silent channel description',
-        playSound: false,
-        styleInformation: DefaultStyleInformation(true, true));
-    var iOSPlatformChannelSpecifics =
-        IOSNotificationDetails(presentSound: false);
-    var platformChannelSpecifics = NotificationDetails(
-        androidPlatformChannelSpecifics, iOSPlatformChannelSpecifics);
-    await flutterLocalNotificationsPlugin.show(0, '<b>silent</b> title',
-        '<b>silent</b> body', platformChannelSpecifics);
-  }
-
-  Future<void> _showSoundUriNotification() async {
-    // this calls a method over a platform channel implemented within the example app to return the Uri for the default
-    // alarm sound and uses as the notification sound
-    String alarmUri = await platform.invokeMethod('getAlarmUri');
-    final x = UriAndroidNotificationSound(alarmUri);
-    var androidPlatformChannelSpecifics = AndroidNotificationDetails(
-        'uri channel id', 'uri channel name', 'uri channel description',
-        sound: x,
-        playSound: true,
-        styleInformation: DefaultStyleInformation(true, true));
-    var iOSPlatformChannelSpecifics =
-        IOSNotificationDetails(presentSound: false);
-    var platformChannelSpecifics = NotificationDetails(
-        androidPlatformChannelSpecifics, iOSPlatformChannelSpecifics);
-    await flutterLocalNotificationsPlugin.show(
-        0, 'uri sound title', 'uri sound body', platformChannelSpecifics);
-  }
-
-  Future<void> _showTimeoutNotification() async {
-    var androidPlatformChannelSpecifics = AndroidNotificationDetails(
-        'silent channel id',
-        'silent channel name',
-        'silent channel description',
-        timeoutAfter: 3000,
-        styleInformation: DefaultStyleInformation(true, true));
-    var iOSPlatformChannelSpecifics =
-        IOSNotificationDetails(presentSound: false);
-    var platformChannelSpecifics = NotificationDetails(
-        androidPlatformChannelSpecifics, iOSPlatformChannelSpecifics);
-    await flutterLocalNotificationsPlugin.show(0, 'timeout notification',
-        'Times out after 3 seconds', platformChannelSpecifics);
-  }
-
-  Future<String> _downloadAndSaveFile(String url, String fileName) async {
-    var directory = await getApplicationDocumentsDirectory();
-    var filePath = '${directory.path}/$fileName';
-    var response = await http.get(url);
-    var file = File(filePath);
-    await file.writeAsBytes(response.bodyBytes);
-    return filePath;
-  }
-
-  Future<void> _showBigPictureNotification() async {
-    var largeIconPath = await _downloadAndSaveFile(
-        'http://via.placeholder.com/48x48', 'largeIcon');
-    var bigPicturePath = await _downloadAndSaveFile(
-        'http://via.placeholder.com/400x800', 'bigPicture');
-    var bigPictureStyleInformation = BigPictureStyleInformation(
-        FilePathAndroidBitmap(bigPicturePath),
-        largeIcon: FilePathAndroidBitmap(largeIconPath),
-        contentTitle: 'overridden <b>big</b> content title',
-        htmlFormatContentTitle: true,
-        summaryText: 'summary <i>text</i>',
-        htmlFormatSummaryText: true);
-    var androidPlatformChannelSpecifics = AndroidNotificationDetails(
-        'big text channel id',
-        'big text channel name',
-        'big text channel description',
-        styleInformation: bigPictureStyleInformation);
-    var platformChannelSpecifics =
-        NotificationDetails(androidPlatformChannelSpecifics, null);
-    await flutterLocalNotificationsPlugin.show(
-        0, 'big text title', 'silent body', platformChannelSpecifics);
-  }
-
-  Future<void> _showBigPictureNotificationHideExpandedLargeIcon() async {
-    var largeIconPath = await _downloadAndSaveFile(
-        'http://via.placeholder.com/48x48', 'largeIcon');
-    var bigPicturePath = await _downloadAndSaveFile(
-        'http://via.placeholder.com/400x800', 'bigPicture');
-    var bigPictureStyleInformation = BigPictureStyleInformation(
-        FilePathAndroidBitmap(bigPicturePath),
-        hideExpandedLargeIcon: true,
-        contentTitle: 'overridden <b>big</b> content title',
-        htmlFormatContentTitle: true,
-        summaryText: 'summary <i>text</i>',
-        htmlFormatSummaryText: true);
-    var androidPlatformChannelSpecifics = AndroidNotificationDetails(
-        'big text channel id',
-        'big text channel name',
-        'big text channel description',
-        largeIcon: FilePathAndroidBitmap(largeIconPath),
-        styleInformation: bigPictureStyleInformation);
-    var platformChannelSpecifics =
-        NotificationDetails(androidPlatformChannelSpecifics, null);
-    await flutterLocalNotificationsPlugin.show(
-        0, 'big text title', 'silent body', platformChannelSpecifics);
-  }
-
-  Future<void> _showNotificationMediaStyle() async {
-    var largeIconPath = await _downloadAndSaveFile(
-        'http://via.placeholder.com/128x128/00FF00/000000', 'largeIcon');
-    var androidPlatformChannelSpecifics = AndroidNotificationDetails(
-      'media channel id',
-      'media channel name',
-      'media channel description',
-      largeIcon: FilePathAndroidBitmap(largeIconPath),
-      styleInformation: MediaStyleInformation(),
-    );
-    var platformChannelSpecifics =
-        NotificationDetails(androidPlatformChannelSpecifics, null);
-    await flutterLocalNotificationsPlugin.show(
-        0, 'notification title', 'notification body', platformChannelSpecifics);
-  }
-
-  Future<void> _showBigTextNotification() async {
-    var bigTextStyleInformation = BigTextStyleInformation(
-        'Lorem <i>ipsum dolor sit</i> amet, consectetur <b>adipiscing elit</b>, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.',
-        htmlFormatBigText: true,
-        contentTitle: 'overridden <b>big</b> content title',
-        htmlFormatContentTitle: true,
-        summaryText: 'summary <i>text</i>',
-        htmlFormatSummaryText: true);
-    var androidPlatformChannelSpecifics = AndroidNotificationDetails(
-        'big text channel id',
-        'big text channel name',
-        'big text channel description',
-        styleInformation: bigTextStyleInformation);
-    var platformChannelSpecifics =
-        NotificationDetails(androidPlatformChannelSpecifics, null);
-    await flutterLocalNotificationsPlugin.show(
-        0, 'big text title', 'silent body', platformChannelSpecifics);
-  }
-
-  Future<void> _showInboxNotification() async {
-    var lines = List<String>();
-    lines.add('line <b>1</b>');
-    lines.add('line <i>2</i>');
-    var inboxStyleInformation = InboxStyleInformation(lines,
-        htmlFormatLines: true,
-        contentTitle: 'overridden <b>inbox</b> context title',
-        htmlFormatContentTitle: true,
-        summaryText: 'summary <i>text</i>',
-        htmlFormatSummaryText: true);
-    var androidPlatformChannelSpecifics = AndroidNotificationDetails(
-        'inbox channel id', 'inboxchannel name', 'inbox channel description',
-        styleInformation: inboxStyleInformation);
-    var platformChannelSpecifics =
-        NotificationDetails(androidPlatformChannelSpecifics, null);
-    await flutterLocalNotificationsPlugin.show(
-        0, 'inbox title', 'inbox body', platformChannelSpecifics);
-  }
-
-  Future<void> _showMessagingNotification() async {
-    // use a platform channel to resolve an Android drawable resource to a URI.
-    // This is NOT part of the notifications plugin. Calls made over this channel is handled by the app
-    String imageUri = await platform.invokeMethod('drawableToUri', 'food');
-    var messages = List<Message>();
-    // First two person objects will use icons that part of the Android app's drawable resources
-    var me = Person(
-      name: 'Me',
-      key: '1',
-      uri: 'tel:1234567890',
-      icon: DrawableResourceAndroidIcon('me'),
-    );
-    var coworker = Person(
-      name: 'Coworker',
-      key: '2',
-      uri: 'tel:9876543210',
-      icon: FlutterBitmapAssetAndroidIcon('icons/coworker.png'),
-    );
-    // download the icon that would be use for the lunch bot person
-    var largeIconPath = await _downloadAndSaveFile(
-        'http://via.placeholder.com/48x48', 'largeIcon');
-    // this person object will use an icon that was downloaded
-    var lunchBot = Person(
-      name: 'Lunch bot',
-      key: 'bot',
-      bot: true,
-      icon: BitmapFilePathAndroidIcon(largeIconPath),
-    );
-    messages.add(Message('Hi', DateTime.now(), null));
-    messages.add(Message(
-        'What\'s up?', DateTime.now().add(Duration(minutes: 5)), coworker));
-    messages.add(Message(
-        'Lunch?', DateTime.now().add(Duration(minutes: 10)), null,
-        dataMimeType: 'image/png', dataUri: imageUri));
-    messages.add(Message('What kind of food would you prefer?',
-        DateTime.now().add(Duration(minutes: 10)), lunchBot));
-    var messagingStyle = MessagingStyleInformation(me,
-        groupConversation: true,
-        conversationTitle: 'Team lunch',
-        htmlFormatContent: true,
-        htmlFormatTitle: true,
-        messages: messages);
-    var androidPlatformChannelSpecifics = AndroidNotificationDetails(
-        'message channel id',
-        'message channel name',
-        'message channel description',
-        category: 'msg',
-        styleInformation: messagingStyle);
-    var platformChannelSpecifics =
-        NotificationDetails(androidPlatformChannelSpecifics, null);
-    await flutterLocalNotificationsPlugin.show(
-        0, 'message title', 'message body', platformChannelSpecifics);
-
-    // wait 10 seconds and add another message to simulate another response
-    await Future.delayed(Duration(seconds: 10), () async {
-      messages.add(
-          Message('Thai', DateTime.now().add(Duration(minutes: 11)), null));
-      await flutterLocalNotificationsPlugin.show(
-          0, 'message title', 'message body', platformChannelSpecifics);
-    });
-  }
-
-  Future<void> _showGroupedNotifications() async {
-    var groupKey = 'com.android.example.WORK_EMAIL';
-    var groupChannelId = 'grouped channel id';
-    var groupChannelName = 'grouped channel name';
-    var groupChannelDescription = 'grouped channel description';
-    // example based on https://developer.android.com/training/notify-user/group.html
-    var firstNotificationAndroidSpecifics = AndroidNotificationDetails(
-        groupChannelId, groupChannelName, groupChannelDescription,
-        importance: Importance.Max,
-        priority: Priority.High,
-        groupKey: groupKey);
-    var firstNotificationPlatformSpecifics =
-        NotificationDetails(firstNotificationAndroidSpecifics, null);
-    await flutterLocalNotificationsPlugin.show(1, 'Alex Faarborg',
-        'You will not believe...', firstNotificationPlatformSpecifics);
-    var secondNotificationAndroidSpecifics = AndroidNotificationDetails(
-        groupChannelId, groupChannelName, groupChannelDescription,
-        importance: Importance.Max,
-        priority: Priority.High,
-        groupKey: groupKey);
-    var secondNotificationPlatformSpecifics =
-        NotificationDetails(secondNotificationAndroidSpecifics, null);
-    await flutterLocalNotificationsPlugin.show(
-        2,
-        'Jeff Chang',
-        'Please join us to celebrate the...',
-        secondNotificationPlatformSpecifics);
-
-    // create the summary notification to support older devices that pre-date Android 7.0 (API level 24).
-    // this is required is regardless of which versions of Android your application is going to support
-    var lines = List<String>();
-    lines.add('Alex Faarborg  Check this out');
-    lines.add('Jeff Chang    Launch Party');
-    var inboxStyleInformation = InboxStyleInformation(lines,
-        contentTitle: '2 messages', summaryText: 'janedoe@example.com');
-    var androidPlatformChannelSpecifics = AndroidNotificationDetails(
-        groupChannelId, groupChannelName, groupChannelDescription,
-        styleInformation: inboxStyleInformation,
-        groupKey: groupKey,
-        setAsGroupSummary: true);
-    var platformChannelSpecifics =
-        NotificationDetails(androidPlatformChannelSpecifics, null);
-    await flutterLocalNotificationsPlugin.show(
-        3, 'Attention', 'Two messages', platformChannelSpecifics);
-  }
-
-  Future<void> _checkPendingNotificationRequests() async {
-    var pendingNotificationRequests =
-        await flutterLocalNotificationsPlugin.pendingNotificationRequests();
-    return showDialog<void>(
-      context: context,
-      builder: (BuildContext context) {
-        return AlertDialog(
-          content: Text(
-              '${pendingNotificationRequests.length} pending notification requests'),
-          actions: [
-            FlatButton(
-              child: Text('OK'),
-              onPressed: () {
-                Navigator.of(context).pop();
-              },
-            ),
-          ],
-        );
-      },
-    );
-  }
-
-  Future<void> _cancelAllNotifications() async {
-    await flutterLocalNotificationsPlugin.cancelAll();
-  }
-
-  Future<void> _showOngoingNotification() async {
-    var androidPlatformChannelSpecifics = AndroidNotificationDetails(
-        'your channel id', 'your channel name', 'your channel description',
-        importance: Importance.Max,
-        priority: Priority.High,
-        ongoing: true,
-        autoCancel: false);
-    var iOSPlatformChannelSpecifics = IOSNotificationDetails();
-    var platformChannelSpecifics = NotificationDetails(
-        androidPlatformChannelSpecifics, iOSPlatformChannelSpecifics);
-    await flutterLocalNotificationsPlugin.show(0, 'ongoing notification title',
-        'ongoing notification body', platformChannelSpecifics);
-  }
-
-  Future<void> _repeatNotification() async {
-    var androidPlatformChannelSpecifics = AndroidNotificationDetails(
-        'repeating channel id',
-        'repeating channel name',
-        'repeating description');
-    var iOSPlatformChannelSpecifics = IOSNotificationDetails();
-    var platformChannelSpecifics = NotificationDetails(
-        androidPlatformChannelSpecifics, iOSPlatformChannelSpecifics);
-    await flutterLocalNotificationsPlugin.periodicallyShow(0, 'repeating title',
-        'repeating body', RepeatInterval.EveryMinute, platformChannelSpecifics);
-  }
-
-  Future<void> _showDailyAtTime() async {
-    var time = Time(10, 0, 0);
-    var androidPlatformChannelSpecifics = AndroidNotificationDetails(
-        'repeatDailyAtTime channel id',
-        'repeatDailyAtTime channel name',
-        'repeatDailyAtTime description');
-    var iOSPlatformChannelSpecifics = IOSNotificationDetails();
-    var platformChannelSpecifics = NotificationDetails(
-        androidPlatformChannelSpecifics, iOSPlatformChannelSpecifics);
-    await flutterLocalNotificationsPlugin.showDailyAtTime(
-        0,
-        'show daily title',
-        'Daily notification shown at approximately ${_toTwoDigitString(time.hour)}:${_toTwoDigitString(time.minute)}:${_toTwoDigitString(time.second)}',
-        time,
-        platformChannelSpecifics);
-  }
-
-  Future<void> _showWeeklyAtDayAndTime() async {
-    var time = Time(10, 0, 0);
-    var androidPlatformChannelSpecifics = AndroidNotificationDetails(
-        'show weekly channel id',
-        'show weekly channel name',
-        'show weekly description');
-    var iOSPlatformChannelSpecifics = IOSNotificationDetails();
-    var platformChannelSpecifics = NotificationDetails(
-        androidPlatformChannelSpecifics, iOSPlatformChannelSpecifics);
-    await flutterLocalNotificationsPlugin.showWeeklyAtDayAndTime(
-        0,
-        'show weekly title',
-        'Weekly notification shown on Monday at approximately ${_toTwoDigitString(time.hour)}:${_toTwoDigitString(time.minute)}:${_toTwoDigitString(time.second)}',
-        Day.Monday,
-        time,
-        platformChannelSpecifics);
-  }
-
-  Future<void> _showNotificationWithNoBadge() async {
-    var androidPlatformChannelSpecifics = AndroidNotificationDetails(
-        'no badge channel', 'no badge name', 'no badge description',
-        channelShowBadge: false,
-        importance: Importance.Max,
-        priority: Priority.High,
-        onlyAlertOnce: true);
-    var iOSPlatformChannelSpecifics = IOSNotificationDetails();
-    var platformChannelSpecifics = NotificationDetails(
-        androidPlatformChannelSpecifics, iOSPlatformChannelSpecifics);
-    await flutterLocalNotificationsPlugin.show(
-        0, 'no badge title', 'no badge body', platformChannelSpecifics,
-        payload: 'item x');
-  }
-
-  Future<void> _showProgressNotification() async {
-    var maxProgress = 5;
-    for (var i = 0; i <= maxProgress; i++) {
-      await Future.delayed(Duration(seconds: 1), () async {
-        var androidPlatformChannelSpecifics = AndroidNotificationDetails(
-            'progress channel',
-            'progress channel',
-            'progress channel description',
-            channelShowBadge: false,
-            importance: Importance.Max,
-            priority: Priority.High,
-            onlyAlertOnce: true,
-            showProgress: true,
-            maxProgress: maxProgress,
-            progress: i);
-        var iOSPlatformChannelSpecifics = IOSNotificationDetails();
-        var platformChannelSpecifics = NotificationDetails(
-            androidPlatformChannelSpecifics, iOSPlatformChannelSpecifics);
-        await flutterLocalNotificationsPlugin.show(
-            0,
-            'progress notification title',
-            'progress notification body',
-            platformChannelSpecifics,
-            payload: 'item x');
-      });
+  tapAction(String payload) {
+    if (payload != null && payload.length > 0) {
+      debugPrint("payload : $payload");
+      List<String> payData = payload.split('-');
+      int typeId = int.parse(payData[0]);
+      int targetId = int.parse(payData[1]);
+      // selectNotificationSubject.add('0-0');
+      handleTapNavAction(typeId, targetId);
     }
   }
 
-  Future<void> _showIndeterminateProgressNotification() async {
-    var androidPlatformChannelSpecifics = AndroidNotificationDetails(
-        'indeterminate progress channel',
-        'indeterminate progress channel',
-        'indeterminate progress channel description',
-        channelShowBadge: false,
-        importance: Importance.Max,
-        priority: Priority.High,
-        onlyAlertOnce: true,
-        showProgress: true,
-        indeterminate: true);
-    var iOSPlatformChannelSpecifics = IOSNotificationDetails();
-    var platformChannelSpecifics = NotificationDetails(
-        androidPlatformChannelSpecifics, iOSPlatformChannelSpecifics);
-    await flutterLocalNotificationsPlugin.show(
-        0,
-        'indeterminate progress notification title',
-        'indeterminate progress notification body',
-        platformChannelSpecifics,
-        payload: 'item x');
-  }
-
-  Future<void> _showNotificationWithUpdatedChannelDescription() async {
-    var androidPlatformChannelSpecifics = AndroidNotificationDetails(
-        'your channel id',
-        'your channel name',
-        'your updated channel description',
-        importance: Importance.Max,
-        priority: Priority.High,
-        channelAction: AndroidNotificationChannelAction.Update);
-    var iOSPlatformChannelSpecifics = IOSNotificationDetails();
-    var platformChannelSpecifics = NotificationDetails(
-        androidPlatformChannelSpecifics, iOSPlatformChannelSpecifics);
-    await flutterLocalNotificationsPlugin.show(
-        0,
-        'updated notification channel',
-        'check settings to see updated channel description',
-        platformChannelSpecifics,
-        payload: 'item x');
-  }
-
-  Future<void> _showPublicNotification() async {
-    var androidPlatformChannelSpecifics = AndroidNotificationDetails(
-        'your channel id', 'your channel name', 'your channel description',
-        importance: Importance.Max,
-        priority: Priority.High,
-        ticker: 'ticker',
-        visibility: NotificationVisibility.Public);
-    var iOSPlatformChannelSpecifics = IOSNotificationDetails();
-    var platformChannelSpecifics = NotificationDetails(
-        androidPlatformChannelSpecifics, iOSPlatformChannelSpecifics);
-    await flutterLocalNotificationsPlugin.show(0, 'public notification title',
-        'public notification body', platformChannelSpecifics,
-        payload: 'item x');
-  }
-
-  Future<void> _showNotificationWithIconBadge() async {
-    var androidPlatformChannelSpecifics = AndroidNotificationDetails(
-        'icon badge channel', 'icon badge name', 'icon badge description');
-    var iOSPlatformChannelSpecifics = IOSNotificationDetails(badgeNumber: 1);
-    var platformChannelSpecifics = NotificationDetails(
-        androidPlatformChannelSpecifics, iOSPlatformChannelSpecifics);
-    await flutterLocalNotificationsPlugin.show(
-        0, 'icon badge title', 'icon badge body', platformChannelSpecifics,
-        payload: 'item x');
-  }
-
-  Future<void> _showNotificationWithoutTimestamp() async {
-    var androidPlatformChannelSpecifics = AndroidNotificationDetails(
-        'your channel id', 'your channel name', 'your channel description',
-        importance: Importance.Max, priority: Priority.High, showWhen: false);
-    var iOSPlatformChannelSpecifics = IOSNotificationDetails();
-    var platformChannelSpecifics = NotificationDetails(
-        androidPlatformChannelSpecifics, iOSPlatformChannelSpecifics);
-    await flutterLocalNotificationsPlugin.show(
-        0, 'plain title', 'plain body', platformChannelSpecifics,
-        payload: 'item x');
-  }
-
-  Future<void> _showNotificationWithAttachment() async {
-    var bigPicturePath = await _downloadAndSaveFile(
-        'http://via.placeholder.com/600x200', 'bigPicture.jpg');
-    var iOSPlatformChannelSpecifics = IOSNotificationDetails(
-        attachments: [IOSNotificationAttachment(bigPicturePath)]);
-    var bigPictureAndroidStyle =
-        BigPictureStyleInformation(FilePathAndroidBitmap(bigPicturePath));
-    var androidPlatformChannelSpecifics = AndroidNotificationDetails(
-        'your channel id', 'your channel name', 'your channel description',
-        importance: Importance.High,
-        priority: Priority.High,
-        styleInformation: bigPictureAndroidStyle);
-    var notificationDetails = NotificationDetails(
-        androidPlatformChannelSpecifics, iOSPlatformChannelSpecifics);
-    await flutterLocalNotificationsPlugin.show(
-        0,
-        'notification with attachment title',
-        'notification with attachment body',
-        notificationDetails);
-  }
-
-  Future<void> _createNotificationChannel() async {
-    var androidNotificationChannel = AndroidNotificationChannel(
-      'your channel id 2',
-      'your channel name 2',
-      'your channel description 2',
-    );
-    await flutterLocalNotificationsPlugin
-        .resolvePlatformSpecificImplementation<
-            AndroidFlutterLocalNotificationsPlugin>()
-        ?.createNotificationChannel(androidNotificationChannel);
-
-    await showDialog<void>(
-        context: context,
-        builder: (BuildContext context) {
-          return AlertDialog(
-            content: Text(
-                'Channel with name \"${androidNotificationChannel.name}\" created'),
-            actions: [
-              FlatButton(
-                child: Text('OK'),
-                onPressed: () {
-                  Navigator.of(context).pop();
-                },
-              ),
-            ],
-          );
+  void handleTapNavAction(int typeId, int targetId) async {
+    // debugPrint('enter ===============>');
+    switch (typeId) {
+      case 0:
+        debugPrint('do not have any notification.');
+        break;
+      case 1:
+        debugPrint('enter to nav to user chat screen===============>');
+        Friend friend = FriendProvider.getFriendById(targetId);
+        msgRepo.getMessagesByRecvId(1, targetId).then((messages) {
+          Navigator.of(_ctx).push(new MaterialPageRoute(builder: (_) {
+            debugPrint('navigation==========================>');
+            return ChatScreen(
+                userType: 1,
+                // username: username,
+                // avatar: avatar,
+                friend: friend,
+                messages: messages);
+          }));
         });
-  }
-
-  Future<void> _deleteNotificationChannel() async {
-    const channelId = 'your channel id 2';
-    await flutterLocalNotificationsPlugin
-        .resolvePlatformSpecificImplementation<
-            AndroidFlutterLocalNotificationsPlugin>()
-        ?.deleteNotificationChannel(channelId);
-
-    await showDialog<void>(
-        context: context,
-        builder: (BuildContext context) {
-          return AlertDialog(
-            content: Text('Channel with id \"$channelId\" deleted'),
-            actions: [
-              FlatButton(
-                child: Text('OK'),
-                onPressed: () {
-                  Navigator.of(context).pop();
-                },
-              ),
-            ],
-          );
-        });
-  }
-
-  String _toTwoDigitString(int value) {
-    return value.toString().padLeft(2, '0');
+        break;
+      case 2:
+        Friend friend = FriendProvider.getFriendById(targetId);
+        // _cancelNotification(targetId);
+        Navigator.of(_ctx).push(new MaterialPageRoute(builder: (_) {
+          return ChatScreen(
+              userType: 1,
+              // username: username,
+              // avatar: avatar,
+              friend: friend,
+              messages: []);
+        }));
+        break;
+      case 3:
+        break;
+    }
   }
 }
