@@ -1,38 +1,42 @@
 import 'package:flutter/material.dart';
-// import 'package:cached_network_image/cached_network_image.dart';
-import 'package:keep/UI/storageBox/storage_box.dart';
+import 'package:keep/UI/Home/Chat/ChatRooms/chat_rooms.dart';
+import 'package:keep/UI/storageBox/storage_page.dart';
 import 'package:keep/data/provider/user_provider.dart';
-import 'package:keep/global/global_tool.dart';
-import 'package:keep/global/user_pic.dart';
+import 'package:keep/utils/tools_function.dart';
+import 'package:keep/widget/user_pic.dart';
 import 'package:provider/provider.dart';
-import 'friends/friends_list_page.dart';
+import 'Chat/Contacts/friends_list_page.dart';
 import 'Settings/setting_main_page.dart';
+import 'Chat/NewGroup/new_group.dart';
+import 'package:keep/settings/selection_config.dart';
 
 class NavDrawer extends StatelessWidget {
-  const NavDrawer({Key key}) : super(key: key);
+  final BuildContext parentContext;
+  const NavDrawer({@required this.parentContext, Key key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
-    return Consumer<UserProvider>(builder: (context, user, child) {
-      return Drawer(
-          child: ListView(
-              shrinkWrap: true,
-              physics: NeverScrollableScrollPhysics(),
-              padding: EdgeInsets.zero,
-              children: <Widget>[
-            DrawerHeader(
-                decoration: BoxDecoration(
-                    image: DecorationImage(
-                        image: AssetImage("assets/images/nav_me.jpg"),
-                        fit: BoxFit.cover)),
-                child: Column(
+    return Drawer(
+        child: ListView(
+            shrinkWrap: true,
+            physics: NeverScrollableScrollPhysics(),
+            padding: EdgeInsets.zero,
+            children: <Widget>[
+          DrawerHeader(
+              decoration: BoxDecoration(
+                  image: DecorationImage(
+                      image: AssetImage("assets/images/nav_me.jpg"),
+                      fit: BoxFit.cover)),
+              child: Consumer<UserProvider>(builder: (context, user, _) {
+                return Column(
                   crossAxisAlignment: CrossAxisAlignment.start,
                   mainAxisAlignment: MainAxisAlignment.spaceBetween,
                   children: <Widget>[
-                    UserAvatar(user.username, user.avatar),
+                    UserAvatar(user.username, user.userAvatar),
                     Container(
                         child: Column(
                       crossAxisAlignment: CrossAxisAlignment.start,
+                      mainAxisAlignment: MainAxisAlignment.spaceEvenly,
                       children: <Widget>[
                         Container(
                             padding: EdgeInsets.symmetric(vertical: 2.0),
@@ -50,75 +54,94 @@ class NavDrawer extends StatelessWidget {
                       ],
                     )),
                   ],
-                )),
-            ListTile(
-              leading: Icon(Icons.group_add),
-              title: Text('New Group'),
-              onTap: () => {},
+                );
+              })),
+          ListTile(
+            leading: Icon(Icons.group_add),
+            title: Text('New Group'),
+            onTap: () async {
+              Navigator.pop(context);
+              Navigator.of(parentContext)
+                  .push(new MaterialPageRoute(builder: (_) {
+                return NewGroupPage();
+              }));
+            },
+          ),
+          ListTile(
+            leading: Icon(Icons.notifications_active),
+            title: Text('Reminders'),
+            trailing: CircleAvatar(
+              radius: 10.0,
+              child: new Text('3',
+                  style: TextStyle(
+                      fontSize: 10,
+                      fontWeight: FontWeight.w600,
+                      color: Colors.white)),
+              backgroundColor: const Color(0xFFf46464),
             ),
-            ListTile(
-              leading: Icon(Icons.notifications_active),
-              title: Text('Notifications'),
-              trailing: CircleAvatar(
-                radius: 10.0,
-                child: new Text('3',
-                    style: TextStyle(
-                        fontSize: 10,
-                        fontWeight: FontWeight.w600,
-                        color: Colors.white)),
-                backgroundColor: const Color(0xFFf46464),
-              ),
-              onTap: () => Navigator.of(context).pop(),
-            ),
-            ListTile(
-                leading: Icon(Icons.collections_bookmark),
-                title: Text('Storage Box'),
-                // trailing: Icon(Icons.keyboard_arrow_right, color: Colors.white),
-                onTap: () => Navigator.of(context)
-                        .push(new MaterialPageRoute(builder: (_) {
-                      return StorageBox(title: 'demo');
-                    }))),
-            ListTile(
-              leading: Icon(Icons.person_outline),
-              title: Text('Contacts'),
-              trailing: Icon(Icons.keyboard_arrow_right, color: Colors.white),
+            onTap: () => Navigator.of(context).pop(),
+          ),
+          ListTile(
+              leading: Icon(Icons.collections_bookmark),
+              title: Text('Storage Box'),
+              // trailing: Icon(Icons.keyboard_arrow_right, color: Colors.white),
               onTap: () {
-                Navigator.of(context).pop();
-                Navigator.of(context).push(new MaterialPageRoute(builder: (_) {
-                  return FriendsListPage();
+                Navigator.pop(context);
+                Navigator.of(parentContext)
+                    .push(new MaterialPageRoute(builder: (_) {
+                  return StoragePage();
                 }));
-              },
-            ),
-            ListTile(
+              }),
+          ListTile(
+            leading: Icon(Icons.person_outline),
+            title: Text('Contacts'),
+            trailing: Icon(Icons.keyboard_arrow_right, color: Colors.white),
+            onTap: () {
+              Navigator.of(parentContext)
+                  .push(new MaterialPageRoute(builder: (_) {
+                return FriendsListPage();
+              }));
+              // Navigator.of(context).pop();
+            },
+          ),
+          ListTile(
               leading: Icon(Icons.group),
               title: Text('ChatRooms'),
               trailing: Icon(Icons.keyboard_arrow_right, color: Colors.white),
-              onTap: () => Navigator.of(context).pop(),
-            ),
-            ListTile(
-              leading: Icon(Icons.feedback),
-              title: Text('Send Feedback'),
-              onTap: () => Navigator.of(context).pop(),
-            ),
-            ListTile(
-              leading: Icon(Icons.settings),
-              title: Text('settings'),
-              // trailing: Icon(Icons.keyboard_arrow_right, color: Colors.white),
               onTap: () {
-                Navigator.of(context)
+                Navigator.of(parentContext)
                     .push(new MaterialPageRoute(builder: (_) {
-                  return SettingMainPage();
+                  return ChatRoomPage();
                 }));
-              },
-            ),
-            Divider(
-              height: 1.0,
-            ),
-            ListTile(
-                leading: Icon(Icons.exit_to_app),
-                title: Text('Log out'),
-                onTap: () => showDialog(
-                    context: context,
+                // Navigator.of(context).pop();
+              }),
+          ListTile(
+            leading: Icon(Icons.feedback),
+            title: Text('Send Feedback'),
+            onTap: () => Navigator.of(context).pop(),
+          ),
+          ListTile(
+            leading: Icon(Icons.settings),
+            title: Text('settings'),
+            // trailing: Icon(Icons.keyboard_arrow_right, color: Colors.white),
+            onTap: () {
+              Navigator.pop(context);
+              Navigator.of(parentContext)
+                  .push(new MaterialPageRoute(builder: (_) {
+                return SettingMainPage();
+              }));
+            },
+          ),
+          Divider(
+            height: 1.0,
+          ),
+          ListTile(
+              leading: Icon(Icons.exit_to_app),
+              title: Text('Log out'),
+              onTap: () {
+                Navigator.pop(context);
+                showDialog(
+                    context: parentContext,
                     child: FractionallySizedBox(
                         widthFactor: 0.85,
                         child: AlertDialog(
@@ -138,17 +161,17 @@ class NavDrawer extends StatelessWidget {
                             FlatButton(
                               onPressed: () {
                                 // print("you choose no");
-                                Navigator.of(context).pop(false);
+                                Navigator.of(parentContext).pop(false);
                               },
                               child: Text('Cancel'),
                             ),
                             FlatButton(
-                              onPressed: logout(context),
+                              onPressed: logout(parentContext),
                               child: Text('Log Out'),
                             ),
                           ],
-                        ))))
-          ]));
-    });
+                        )));
+              })
+        ]));
   }
 }
