@@ -1,19 +1,24 @@
-import 'package:keep/global/global_tool.dart';
+import 'package:keep/utils/tools_function.dart';
 
 class User {
   int userId;
   String email;
   String password;
   String username;
+  String phone;
   String apiKey;
   String userPic;
+  Object avatar;
   User(
       {this.userId,
       this.email,
       this.password,
       this.username,
       this.apiKey,
-      this.userPic});
+      String base64Txt,
+      this.phone = 'do not set'}) : 
+      this.userPic = base64Txt,
+      this.avatar = base64Txt == 'null' ? 'null' : txt2Image(base64Txt);
 
   User.map(dynamic obj) {
     this.userId = int.parse(obj['user_id']);
@@ -22,6 +27,7 @@ class User {
     this.username = obj["username"];
     this.apiKey = obj["api_key"] ?? "You don't know";
     this.userPic = obj["user_pic"];
+    this.phone = obj["phone"] ?? 'do not set';
   }
 
   static User fromMap(Map map) {
@@ -30,8 +36,9 @@ class User {
         email: map['email'],
         password: map['password'] ?? 'You do not know',
         username: capitalize(map['username']),
+        phone: map['phone'] ?? 'do not set',
         apiKey: map['api_key'],
-        userPic: map['user_pic']);
+        base64Txt: map['user_pic']);
   }
 
   Map toJson() {
@@ -41,6 +48,7 @@ class User {
       'email': email,
       'password': password,
       'username': username,
+      'phone': phone,
       'api_key': apiKey,
       'user_pic': userPic
     };
@@ -52,7 +60,8 @@ class User {
         email: json['email'] as String,
         password: json['password'] as String,
         username: json['username'] as String,
+        phone: json['phone'] as String,
         apiKey: json['api_key'] as String,
-        userPic: json['user_pic'] as String);
+        base64Txt: json['user_pic'] as String);
   }
 }
